Refactor friends list state so it's more compatible with sectioned lists
Previously the friends list state contained the sorted and sectioned friends
list in a flat array since this was compatible with NSTableView. Now it contains
the state in what is effectively a nested array of sections. This is more
compatible with iOS/watchOS/SwiftUI sectioning. The transformation to the flat
list needed for AppKit is done closer to the AppKit layer now since it's more
AppKit specific.